requirements.txt
streamlit==0.79.0
Pillow==8.1.2

